client: fix parse_chunksize() return type and stale comment
parse_chunksize() used size_t as its return type while relying on -1
as an error sentinel, which only works because size_t and long happen
to be the same width on our build targets; use long instead so the
sentinel is unambiguous. Also update a leftover comment that still
described strtol()'s empty-string behavior even though strtol() is no
longer used for chunk size parsing.

+++ client.c
@@ -396,7 +396,7 @@ static uint8_t chartypes[256] = {
 /* 80..ff: no flags */
 };
 
-static size_t
+static long
 parse_chunksize(char *buf, char *end)
 {
 	int size = 0;
@@ -611,10 +611,9 @@ void client_poll_post_data(struct client *cl)
 
 		chunk_len = parse_chunksize(buf + offset, sep);
 
-		/* invalid chunk length, including an empty chunk size line
-		 * (strtol returns 0 with sep pointing to the start of the
-		 * string when no hex digits are present, which would be
-		 * indistinguishable from a valid 0-size final chunk) */
+		/* invalid chunk length, including an empty chunk size line,
+		 * leading/trailing garbage or an overflowing value; see
+		 * parse_chunksize() */
 		if (chunk_len < 0) {
 			ustream_consume(cl->us, sep + 2 - buf);
 			r->content_length = 0;
